requirements.txt
scikit-learn==1.3.2
seaborn==0.13.1
numpy==1.26.4
matplotlib==3.7.1
optuna==3.6.1
PyYAML>=6.0
lightgbm
polars
duckdb
pandas
ipywidgets
jupyterlab
jupyter
plotly
umap-learn
fastparquet

google-cloud-bigquery
google-cloud-storage
google-cloud-bigquery-storage
pyarrow
